# openapi-appengine.yaml
...
swagger: "2.0"
info:
  description: "A simple Google Cloud Endpoints API example."
  title: "Endpoints Example"
  version: "1.0.0"
host: "nodejs-server-273711.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
paths:
...
  "/user/getProfile":
    post:
      description: "Get profile for user"
      operationId: "getProfile"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/GetProfileParams"
      responses:
        200:
          description: "Response for get profile"
          schema:
            $ref: "#/definitions/GetProfileResponse"
  "/blog/mainPosts":
    post:
      description: "Get main post"
      operationId: "getMainPosts"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/GetMainPostsParams"
      responses:
        200:
          description: "get main posts"
          schema:
            $ref: "#/definitions/GetMainPostsResponse"
...
definitions:
...
  GetProfileParams:
    type: "object"
    properties:
      userId:
        type: "string"
  GetProfileResponse:
    type: "object"
    properties:
      Id:
        type: "string"
      UserName:
        type: "string"
      FbUrl:
        type: "string"
      TechTxt:
        type: "string"
      AddTxt:
        type: "string"
  GetMainPostsParams:
    type: "object"
  GetMainPostsResponse:
    type: "object"
    properties:
      Id:
        type: "string"
      Author:
        type: "string"
      Title:
        type: "string"
      Content:
        type: "string"
      Topic:
        type: "string"
      SubmitDate:
        type: "string"
      ImageUrl:
        type: "string"
...
security:
  - api_key: []
securityDefinitions:
  # This section configures basic authentication with an API key.
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"